at new time the fx app hasn't even started so shutdowner doesn't work
