+++ athenareader/configfx/module.go
@@ -21,15 +21,14 @@
 package configfx
 
 import (
+	"context"
 	"flag"
 	"fmt"
 	secret "github.com/uber/athenadriver/examples/constants"
 	drv "github.com/uber/athenadriver/go"
 	"go.uber.org/config"
 	"go.uber.org/fx"
-	"io"
 	"io/ioutil"
-	"net/http"
 	"os"
 )
 
@@ -39,7 +38,8 @@ var Module = fx.Provide(new)
 // Params defines the dependencies or inputs
 type Params struct {
 	fx.In
-	Shutdowner fx.Shutdowner
+
+	LC fx.Lifecycle
 }
 
 // ReaderInputConfig is to represent the output section of configuration file
@@ -89,50 +89,26 @@ type Result struct {
 }
 
 func init() {
-	os.Setenv("AWS_SDK_LOAD_CONFIG", "1")
-	var commandLine = flag.NewFlagSet(os.Args[0], flag.ExitOnError)
-	flag.Usage = func() {
-		preBody := "NAME\n\tathenareader - read athena data from command line\n\n"
-		desc := "\nEXAMPLES\n\n" +
-			"\t$ athenareader -d sampledb -q \"select request_timestamp,elb_name from elb_logs limit 2\"\n" +
-			"\trequest_timestamp,elb_name\n" +
-			"\t2015-01-03T00:00:00.516940Z,elb_demo_004\n" +
-			"\t2015-01-03T00:00:00.902953Z,elb_demo_004\n\n" +
-			"\t$ athenareader -d sampledb -q \"select request_timestamp,elb_name from elb_logs limit 2\" -r\n" +
-			"\t2015-01-05T20:00:01.206255Z,elb_demo_002\n" +
-			"\t2015-01-05T20:00:01.612598Z,elb_demo_008\n\n" +
-			"\t$ athenareader -d sampledb -b s3://my-athena-query-result -q tools/query.sql\n" +
-			"\trequest_timestamp,elb_name\n" +
-			"\t2015-01-06T00:00:00.516940Z,elb_demo_009\n\n" +
-			"\n\tAdd '-m' to enable moneywise mode. The first line will display query cost under moneywise mode.\n\n" +
-			"\t$ athenareader -b s3://athena-query-result -q 'select count(*) as cnt from sampledb.elb_logs' -m\n" +
-			"\tquery cost: 0.00184898369752772851 USD\n" +
-			"\tcnt\n" +
-			"\t1356206\n\n" +
-			"\n\tAdd '-a' to enable admin mode. Database write is enabled at driver level under admin mode.\n\n" +
-			"\t$ athenareader -b s3://athena-query-result -q 'DROP TABLE IF EXISTS depreacted_table' -a\n" +
-			"\t\n" +
-			"AUTHOR\n\tHenry Fuheng Wu (henry.wu@uber.com)\n\n" +
-			"REPORTING BUGS\n\thttps://github.com/uber/athenadriver\n"
-		fmt.Fprintf(commandLine.Output(), preBody)
-		fmt.Fprintf(commandLine.Output(),
-			"SYNOPSIS\n\n\t%s [-v] [-b OUTPUT_BUCKET] [-d DATABASE_NAME] [-q QUERY_STRING_OR_FILE] [-r] [-a] [-m] [-y STYLE_NAME] [-o OUTPUT_FORMAT]\n\nDESCRIPTION\n\n", os.Args[0])
-		flag.PrintDefaults()
-		fmt.Fprintf(commandLine.Output(), desc)
-	}
-}
-
-func isFlagPassed(name string) bool {
-	found := false
-	flag.Visit(func(f *flag.Flag) {
-		if f.Name == name {
-			found = true
-		}
-	})
-	return found
+	setUpFlagUsage(context.Background())
 }
 
 func new(p Params) (Result, error) {
+	var mc = AthenaDriverConfig{}
+	var (
+		provider *config.YAML
+		err      error
+	)
+
+	p.LC.Append(fx.Hook{
+		OnStart: func(ctx context.Context) error {
+			return nil
+		},
+		OnStop: func(ctx context.Context) error {
+			os.Unsetenv("AWS_SDK_LOAD_CONFIG")
+			return nil
+		},
+	})
+
 	var bucket = flag.String("b", secret.OutputBucket, "Athena resultset output bucket")
 	var database = flag.String("d", "default", "The database you want to query")
 	var query = flag.String("q", "select 1", "The SQL query string or a file containing SQL string")
@@ -147,16 +123,9 @@ func new(p Params) (Result, error) {
 	switch {
 	case *versionFlag:
 		println("Current build version: v" + drv.DriverVersion)
-		p.Shutdowner.Shutdown()
 		os.Exit(0)
 		return Result{}, fmt.Errorf("no")
 	}
-
-	var mc = AthenaDriverConfig{}
-	var (
-		provider *config.YAML
-		err      error
-	)
 
 	if _, err = os.Stat(homeDir() + "/athenareader.config"); err == nil {
 		provider, err = config.NewYAML(config.File(homeDir() + "/athenareader.config"))
@@ -169,8 +138,7 @@ func new(p Params) (Result, error) {
 			if _, err = os.Stat(goPath); err != nil {
 				d, _ := os.Getwd()
 				println("could not find athenareader.config in home directory or current directory " + d)
-				p.Shutdowner.Shutdown()
-				os.Exit(2)
+				os.Exit(1)
 			}
 		}
 		path := goPath + "/src/github.com/uber/athenadriver/athenareader/athenareader.config"
@@ -183,8 +151,7 @@ func new(p Params) (Result, error) {
 			if err != nil {
 				d, _ := os.Getwd()
 				println("could not find athenareader.config in home directory or current directory " + d)
-				p.Shutdowner.Shutdown()
-				os.Exit(2)
+				os.Exit(1)
 			} else {
 				provider, err = config.NewYAML(config.File(homeDir() + "/athenareader.config"))
 			}
@@ -245,56 +212,4 @@ func new(p Params) (Result, error) {
 	return Result{
 		MyConfig: mc,
 	}, nil
-}
-
-func copyFile(src, dst string) error {
-	in, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer in.Close()
-
-	out, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-	defer out.Close()
-
-	_, err = io.Copy(out, in)
-	if err != nil {
-		return err
-	}
-	return out.Close()
-}
-
-func homeDir() string {
-	if h := os.Getenv("HOME"); h != "" {
-		return h
-	}
-	return os.Getenv("USERPROFILE") // windows
-}
-
-func downloadFile(filepath string, url string) (err error) {
-	// Create the file
-	out, err := os.Create(filepath)
-	if err != nil {
-		return err
-	}
-	defer out.Close()
-	// Get the data
-	resp, err := http.Get(url)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	// Check server response
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("bad status: %s", resp.Status)
-	}
-	// Writer the body to file
-	_, err = io.Copy(out, resp.Body)
-	if err != nil {
-		return err
-	}
-	return nil
 }
